Remove redundant code from Conf subclasses to base, and remove references to Ryu from meter Conf.
