+++ faucet/port.py
@@ -83,16 +83,11 @@ class Port(Conf):
     }
 
     def __init__(self, _id, conf=None):
-        if conf is None:
-            conf = {}
-        self._id = _id
-        self.update(conf)
-        self.set_defaults()
+        super(Port, self).__init__(_id, conf)
         self.dyn_phys_up = False
 
     def set_defaults(self):
-        for key, value in list(self.defaults.items()):
-            self._set_default(key, value)
+        super(Port, self).__set_defaults()
         self._set_default('number', self._id)
         self._set_default('name', str(self._id))
         self._set_default('description', self.name)
