+++ faucet/vlan.py
@@ -111,12 +111,8 @@ class VLAN(Conf):
     }
 
     def __init__(self, _id, dp_id, conf=None):
-        if conf is None:
-            conf = {}
-        self._id = _id
+        super(VLAN, self).__init__(_id, conf)
         self.dp_id = dp_id
-        self.update(conf)
-        self.set_defaults()
         self.tagged = []
         self.untagged = []
         self.dyn_host_cache = {}
@@ -180,8 +176,7 @@ class VLAN(Conf):
         self.dyn_host_cache = value
 
     def set_defaults(self):
-        for key, value in list(self.defaults.items()):
-            self._set_default(key, value)
+        super(VLAN, self).set_defaults()
         self._set_default('vid', self._id)
         self._set_default('name', str(self._id))
         self._set_default('faucet_vips', [])
